replace int fields manual lock with atomic lock and remove extra clone

+++ feature/s3/transfermanager/download_directory_test.go
@@ -59,8 +59,8 @@ func TestDownloadDirectory(t *testing.T) {
 		expectKeys              []string
 		expectFiles             []string
 		expectErr               string
-		expectObjectsDownloaded int
-		expectObjectsFailed     int
+		expectObjectsDownloaded int64
+		expectObjectsFailed     int64
 		listenerValidationFn    func(*testing.T, *mockDirectoryListener, any, any, error)
 	}{
 		"single object": {
